har-risk-codex-package/har-risk-codex-package/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "har-risk-engine"
version = "0.1.0"
description = "Evidence-first HAR analysis engine for class action risk signals"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "mypy>=1.10",
  "ruff>=0.5.0",
]

[project.scripts]
har-risk = "har_engine.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["har_engine"]
